+++ pylinac/vmat.py
@@ -120,7 +120,7 @@ class Segment(Rectangle):
 
     @property
     def stdev(self) -> float:
-        """Return the standard deviation of the segment as a percentage of R_corr."""
+        """Return the standard deviation of the segment."""
         dmlc_value = self._dmlc_image.array[
             self.bl_corner.y : self.bl_corner.y + self.height,
             self.bl_corner.x : self.bl_corner.x + self.width,
@@ -130,7 +130,7 @@ class Segment(Rectangle):
             self.bl_corner.x : self.bl_corner.x + self.width,
         ]
         # we multiply by 100 to be consistent w/ r_corr. I.e. this is a % value.
-        return np.std(dmlc_value / open_value) * 100
+        return float(np.std(dmlc_value / open_value))
 
     @property
     def passed(self) -> bool: